Move `BuildHistorySummary` to the `Wallet` class
