+++ WalletWasabi.Fluent/Models/Wallets/WalletTransactionsModel.cs
@@ -43,7 +43,7 @@ public partial class WalletTransactionsModel : ReactiveObject
 
 	public async Task<TransactionSummary?> GetById(string transactionId)
 	{
-		var txRecordList = await Task.Run(() => TransactionHistoryBuilder.BuildHistorySummary(_wallet));
+		var txRecordList = await Task.Run(() => _wallet.BuildHistorySummary());
 
 		var transaction = txRecordList.FirstOrDefault(x => x.GetHash().ToString() == transactionId);
 		return transaction;
@@ -59,6 +59,6 @@ public partial class WalletTransactionsModel : ReactiveObject
 
 	private IEnumerable<TransactionSummary> BuildSummary()
 	{
-		return TransactionHistoryBuilder.BuildHistorySummary(_wallet);
+		return _wallet.BuildHistorySummary();
 	}
 }
