--- WalletWasabi/Blockchain/Transactions/TransactionHistoryBuilder.cs
@@ -1,44 +1,0 @@
-using NBitcoin;
-using System.Collections.Generic;
-using System.Linq;
-using WalletWasabi.Blockchain.TransactionOutputs;
-using WalletWasabi.Extensions;
-using WalletWasabi.Wallets;
-
-namespace WalletWasabi.Blockchain.Transactions;
-
-public class TransactionHistoryBuilder
-{
-	public static List<TransactionSummary> BuildHistorySummary(Wallet wallet)
-	{
-		Dictionary<uint256, TransactionSummary> mapByTxid = new();
-
-		foreach (SmartCoin coin in wallet.GetAllCoins())
-		{
-			if (mapByTxid.TryGetValue(coin.TransactionId, out TransactionSummary? found)) // If found then update.
-			{
-				found.Amount += coin.Amount;
-			}
-			else
-			{
-				mapByTxid.Add(coin.TransactionId, new TransactionSummary(coin.Transaction, coin.Amount));
-			}
-
-			if (coin.SpenderTransaction is { } spenderTransaction)
-			{
-				var spenderTxId = spenderTransaction.GetHash();
-
-				if (mapByTxid.TryGetValue(spenderTxId, out TransactionSummary? foundSpenderCoin)) // If found then update.
-				{
-					foundSpenderCoin.Amount -= coin.Amount;
-				}
-				else
-				{
-					mapByTxid.Add(spenderTxId, new TransactionSummary(spenderTransaction, Money.Zero - coin.Amount));
-				}
-			}
-		}
-
-		return mapByTxid.Values.OrderByBlockchain().ToList();
-	}
-}
